Get following only when we know the account_id
The code used to call mastodon_verify_credentials and immediately
after that it would call mastodon_following. But mastodon_following
just returns if the account_id has not been set and that value is only
set in the mastodon_verify_credentials callback,
mastodon_http_verify_credentials! So now we call mastodon_following
from mastodon_http_verify_credentials and the account_id should always
be set.

Also, fix a bug in mastodon_http_verify_credentials: the account_id is
not simply an int, it's a int64 and sent as a string. This resulted in
the bug of id "12540" being stored as id 5 instead of 12540. With the
wrong id, the wrong followers were being added.