Use list of mentions for MASTODON_MAYBE_REPLY
We do get a list of mentions for the MASTODON_REPLY case but when
handling MASTODON_MAYBE_REPLY, the list of mentions is always empty
and we need to look up the toot we're replying to via
mastodon_user_data. It will also contain the actual list of mentions
we need to use.