Fixed mentions for boosts
When A boosts a toot by B mentioning account C and D, what do you
expect when you're replying to A? Here's what's supposed to happen:
all are mentioned in your reply. A is mentioned because that's the
person you're replying to (this is different in the web app!), B is
mentioned because they're the ones being boosted, and C and D are
mentioned because that's what we do: keep mentioning people.