Remove CW1 but keep spoiler when posting
The code to support CW1 was getting to unwieldy. I think we need a
different solution, here. Something generic which doesn't depend on
client support. I think that would be a separate spoiler command which
allows us to store a spoiler for the next post.

We now keep a copy of the spoiler text for every message in our log,
and for every user we have data on, so that when replying to them, we
can reuse this just as we're reusing visibility and mentions.