Redo of replies with non-public visibility (WIP)
In order to redo replies, we need to store both the CW command and the
REPLY command in the REDO list. These are separated by the ASCII Field
Separator (FS) character and stored in the same position.

When parsing a status we also need to keep a copy of the status
without content warning because the prefix will be reappended.