Replies and visibility and a whole lot of changes
Add VISIBILITY command. The PUBLIC, UNLISTED, PRIVATE and DIRECT
commands are now an error to use.

mastodon_post_message now takes into account the mastodon_data (md)
visibility which we can set using the VISIBILITY command.

The HISTORY command now splits multiple commands correctly.

When posting with a non-default visibility, show which visibility was
used.

mastodon_status_show_chat1 now takes ic parameter because it needs the
default visibility. mastodon_default_visibility moved to lib.
mastodon_visibility is exported because we need it for feedback
mastodon_handle_command ("Next command is direct").

Refactoring: rename the argument status to ms for consistency in a few
places.

Simplified creation of todo in mastodon_http_callback by switching to
a GString.

mastodon_length_check reports an error if the message is empty even if
a CW is given.