Fix mastodon_string_join
The code used to return NULL if the initial list is NULL. This makes
sense when replying to a message by ourselves with no mentions:
mastodon_string_join(mentions, NULL). But when we now reply to the
message of another user who only mentions us, and thus ms->mentions is
empty (because we skip our own username in mastodon_xt_get_status),
then we have a problem: mentions is empty but the init argument is
not. In fact, the calling code already knows that the init argument
cannot be empty and doesn't even check if the returned GString is NULL
or not, it just gets used -- and bitlbee crashes. Thus, we can only
return NULL when both arguments, mentions and init, are NULL.