Added 'list reload' command
For now, a manual 'list reload' is needed after editing lists and list
membership. Toots from people on lists should get distributed to
channels having the same topic as these lists. These statuses are then
marked as seen and should not show up on the normal timeline. Not sure
this is expected behaviour.