Restructure how lists are loaded
When joining a group chat that is supported by a list, we'd like to
populate it with a bunch of messages. After the initial connect, this
practically never works, however: before we can do that, we need to
load the list members, and this only work if we know the members, and
we only know them after having loaded all the followers, and these can
come in multiple pages. That's why some reorganizing was required.