Load list timeline on manual chat join
Set the MASTODON_HAVE_FRIENDS as late as possible: Not as soon as we
start populating the following list but when we have come to the end.
This allows us to determine whether we are joining a group chat
because of auto_join (in which case there's no point in populating the
timeline) or because of a manual join. If due to a manual join, call
mastodon_unknown_list_timeline().