Fixed 'timeline' command
The 'timeline' and 'timeline home' commands were supposed to show the
home timeline, not the user's own toots. So that was implemented. In
addition to that, it's now impossible to know what to do about
'timeline foo' -- is the user interested in the account @foo or in the
list foo? As we don't really know the lists, both questions need a
request to be fulfilled. So now we require a leading '@' and that gets
treated as an account instead of a list. 'timeline @foo' is a user.
'timeline foo' is a list.