Don't rely on md->user
I've never had a problem: my nick is always my Mastodon account
name (account->acct). But for some people, this is not true. In these
cases, replies always mention the account doing the reply because the
plugin doesn't recognize that it's "them" and that they don't need to
be mentioned. This commit fixes it: whenever we need to know whether
the user is involved, we don't compare md->user and ms->account->acct
or whatever it is. Instead, we use the account_id setting we store
after the login and compare that to the ms->account->id.