mastodon_status_show_msg handles self-messages
Right now a reply to a direct message generates a self-message. This
self-message is "from" ourselves, but there is no target for "us" so
the message ends up in the control channel (&bitlbee). This commit
makes sure that if we reply, all the message is delivered once for
every account we mentioned (the recipients), if and only if we find
the appropriate user. This supposedly always works because we're
replying to somebody.

Things to watch out for, however: If we're sending a direct message to
somebody we're not following, in other words we send it from the
account channel, then the self-message has a mention that isn't found
as buddy. What happens?