Switch search URL to /api/v2/search
In order to mix API calls from v1 and v2, we need to add the API
version to the various URL constants, which means that we must no
longer add it to the base_url, or ignore it if a path is provided.
This is a breaking change for instances that aren't situated at the
top level but I'm not even sure whether it is possible to install
Mastodon in such a way.