Fixing login for Pleroma
In addition to using the Authorization HTTP header, we're sending the
authorization token as a parameter, even though I think that's wrong.
But if you do it, it works... Apparently that's how MastoFE
communicates with the backend, too.

We're also ignoring 0 length messages. I wonder what that is all
about.

Sadly we're still getting disconnected every two minutes.